Feature c7 ncd 2450

See merge request choerodon-c7ncd/devops-service!551

+++ src/main/java/io/choerodon/devops/infra/persistence/impl/GitlabRepositoryImpl.java
@@ -3,18 +3,14 @@ package io.choerodon.devops.infra.persistence.impl;
 import java.util.ArrayList;
 import java.util.List;
 
-import feign.FeignException;
-import feign.RetryableException;
 import io.choerodon.core.convertor.ConvertHelper;
 import io.choerodon.core.exception.CommonException;
 import io.choerodon.devops.api.dto.gitlab.MemberDTO;
 import io.choerodon.devops.api.dto.gitlab.VariableDTO;
 import io.choerodon.devops.domain.application.entity.DevopsProjectE;
-
 import io.choerodon.devops.domain.application.repository.GitlabRepository;
 import io.choerodon.devops.domain.application.valueobject.DeployKey;
 import io.choerodon.devops.domain.application.valueobject.ProjectHook;
-import io.choerodon.devops.domain.application.valueobject.RepositoryFile;
 import io.choerodon.devops.domain.application.valueobject.Variable;
 import io.choerodon.devops.infra.common.util.GitUtil;
 import io.choerodon.devops.infra.dataobject.gitlab.GitlabProjectDO;
@@ -55,14 +51,13 @@ public class GitlabRepositoryImpl implements GitlabRepository {
 
     @Override
     public List<String> listTokenByUserId(Integer gitlabProjectId, String name, Integer userId) {
-        ResponseEntity<List<ImpersonationTokenDO>> impersonationTokens;
-        try {
-            impersonationTokens = gitlabServiceClient
-                    .listTokenByUserId(userId);
-        } catch (FeignException e) {
+
+        ResponseEntity<List<ImpersonationTokenDO>> impersonationTokens = gitlabServiceClient
+                .listTokenByUserId(userId);
+        if (impersonationTokens.getStatusCodeValue() == 500) {
             gitUtil.deleteWorkingDirectory(name);
             gitlabServiceClient.deleteProject(gitlabProjectId, userId);
-            throw new CommonException(e);
+            throw new CommonException("error.token.query");
         }
         List<String> tokens = new ArrayList<>();
         impersonationTokens.getBody().stream().forEach(impersonationToken ->
@@ -73,130 +68,75 @@ public class GitlabRepositoryImpl implements GitlabRepository {
 
     @Override
     public String createToken(Integer gitlabProjectId, String name, Integer userId) {
-        ResponseEntity<ImpersonationTokenDO> impersonationToken;
-        try {
-            impersonationToken = gitlabServiceClient.createToken(userId);
-        } catch (FeignException e) {
+        ResponseEntity<ImpersonationTokenDO> impersonationToken = gitlabServiceClient.createToken(userId);
+        if (impersonationToken.getStatusCodeValue() == 500) {
             gitUtil.deleteWorkingDirectory(name);
             gitlabServiceClient.deleteProject(gitlabProjectId, userId);
-            throw new CommonException(e);
+            throw new CommonException("error.token.create");
         }
         return impersonationToken.getBody().getToken();
     }
 
     @Override
     public DevopsProjectE queryGroupByName(String groupName, Integer userId) {
-        ResponseEntity<GroupDO> groupDO;
-        try {
-            groupDO = gitlabServiceClient.queryGroupByName(groupName, userId);
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
-        if (groupDO != null) {
-            return ConvertHelper.convert(groupDO.getBody(), DevopsProjectE.class);
-        } else {
-            return null;
-        }
+
+        ResponseEntity<GroupDO> groupDO = gitlabServiceClient.queryGroupByName(groupName, userId);
+        return ConvertHelper.convert(groupDO.getBody(), DevopsProjectE.class);
+
     }
 
     @Override
     public DevopsProjectE createGroup(DevopsProjectE devopsProjectE, Integer userId) {
-        ResponseEntity<GroupDO> groupDOResponseEntity;
         GroupDO groupDO = ConvertHelper.convert(devopsProjectE, GroupDO.class);
-        try {
-            groupDOResponseEntity = gitlabServiceClient.createGroup(groupDO, userId);
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
+        ResponseEntity<GroupDO> groupDOResponseEntity = gitlabServiceClient.createGroup(groupDO, userId);
         return ConvertHelper.convert(groupDOResponseEntity.getBody(), DevopsProjectE.class);
     }
 
     @Override
     public void createFile(Integer projectId, String path, String content, String commitMessage, Integer userId) {
-        try {
-            ResponseEntity<RepositoryFile> result = gitlabServiceClient
-                    .createFile(projectId, path, content, commitMessage, userId);
-            if (result.getBody().getFilePath() == null) {
-                throw new CommonException("error.file.create");
-            }
-        }catch (RetryableException e) {
-            LOGGER.info(e.getMessage(),e);
-        }
+        gitlabServiceClient
+                .createFile(projectId, path, content, commitMessage, userId);
     }
 
 
     @Override
     public void createFile(Integer projectId, String path, String content, String commitMessage, Integer userId, String branch) {
-        try {
-            ResponseEntity<RepositoryFile> result = gitlabServiceClient
-                    .createFile(projectId, path, content, commitMessage, userId, branch);
-            if (result.getBody().getFilePath() == null) {
-                throw new CommonException("error.file.create");
-            }
-        }catch (RetryableException e) {
-            LOGGER.info(e.getMessage(),e);
-        }
+        gitlabServiceClient
+                .createFile(projectId, path, content, commitMessage, userId, branch);
     }
 
     @Override
     public void updateFile(Integer projectId, String path, String content, String commitMessage, Integer userId) {
-        try {
-            ResponseEntity<RepositoryFile> result = gitlabServiceClient
-                    .updateFile(projectId, path, content, commitMessage, userId);
-            if (result.getBody().getFilePath() == null) {
-                throw new CommonException("error.file.update");
-            }
-        }catch (RetryableException e) {
-            LOGGER.info(e.getMessage(),e);
-        }
+        gitlabServiceClient
+                .updateFile(projectId, path, content, commitMessage, userId);
     }
 
     @Override
     public void deleteFile(Integer projectId, String path, String commitMessage, Integer userId) {
-        try {
-            gitlabServiceClient.deleteFile(projectId, path, commitMessage, userId);
-        } catch (FeignException e) {
-            throw new CommonException("error.file.delete", e);
-        }
+        gitlabServiceClient.deleteFile(projectId, path, commitMessage, userId);
     }
 
     @Override
     public void deleteDevOpsApp(String groupName, String projectName, Integer userId) {
-        try {
-            gitlabServiceClient.deleteProjectByProjectName(groupName, projectName, userId);
-        } catch (FeignException e) {
-            throw new CommonException("error.app.delete", e);
-        }
+        gitlabServiceClient.deleteProjectByProjectName(groupName, projectName, userId);
     }
 
     @Override
     public Boolean getFile(Integer projectId, String branch, String filePath) {
-        try {
-            gitlabServiceClient.getFile(projectId, branch, filePath);
-        } catch (FeignException e) {
-            return false;
-        }
-        return true;
+        ResponseEntity responseEntity = gitlabServiceClient.getFile(projectId, branch, filePath);
+        return responseEntity.getStatusCodeValue() != 500;
     }
 
     @Override
     public void createProtectBranch(Integer projectId, String name, String mergeAccessLevel, String pushAccessLevel,
                                     Integer userId) {
-        try {
             gitlabServiceClient.createProtectedBranches(
                     projectId, name, mergeAccessLevel, pushAccessLevel, userId);
-        } catch (FeignException e) {
-            throw new CommonException("error.branch.create", e);
-        }
     }
 
     @Override
     public void deleteProject(Integer projectId, Integer userId) {
-        try {
-            gitlabServiceClient.deleteProject(projectId, userId);
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
+        gitlabServiceClient.deleteProject(projectId, userId);
     }
 
     @Override
@@ -207,134 +147,76 @@ public class GitlabRepositoryImpl implements GitlabRepository {
 
     @Override
     public ProjectHook createWebHook(Integer projectId, Integer userId, ProjectHook projectHook) {
-        try {
             return gitlabServiceClient.createProjectHook(projectId, userId, projectHook).getBody();
-        } catch (FeignException e) {
-            throw new CommonException("error.projecthook.create", e);
-
-        }
     }
 
     @Override
     public ProjectHook updateWebHook(Integer projectId, Integer hookId, Integer userId) {
-        ResponseEntity<ProjectHook> projectHookResponseEntity;
-        try {
-            projectHookResponseEntity = gitlabServiceClient
+        ResponseEntity<ProjectHook> projectHookResponseEntity = gitlabServiceClient
                     .updateProjectHook(projectId, hookId, userId);
-        } catch (FeignException e) {
-            throw new CommonException(e.getMessage(), e);
-        }
         return projectHookResponseEntity.getBody();
     }
 
     @Override
     public GitlabProjectDO createProject(Integer groupId, String projectName, Integer userId, boolean visibility) {
-        try {
-            return gitlabServiceClient
-                    .createProject(groupId, projectName, userId, visibility).getBody();
-        } catch (FeignException e) {
-            throw new CommonException("error.gitlab.project.create", e);
-
-        }
+        return gitlabServiceClient
+                .createProject(groupId, projectName, userId, visibility).getBody();
     }
 
     @Override
     public GitlabProjectDO getProjectById(Integer projectId) {
-        try {
-            return gitlabServiceClient.getProjectById(projectId).getBody();
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
+        return gitlabServiceClient.getProjectById(projectId).getBody();
     }
 
     @Override
     public GitlabProjectDO getProjectByName(String groupName, String projectName, Integer userId) {
-        try {
-            return gitlabServiceClient.getProjectByName(userId, groupName, projectName).getBody();
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
+        return gitlabServiceClient.getProjectByName(userId, groupName, projectName).getBody();
     }
 
     @Override
     public List<ProjectHook> getHooks(Integer projectId, Integer userId) {
-        try {
             return gitlabServiceClient.getProjectHook(projectId, userId).getBody();
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
     }
 
     @Override
     public List<Variable> getVariable(Integer projectId, Integer userId) {
-        try {
-            return gitlabServiceClient.getVariable(projectId, userId).getBody();
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
+        return gitlabServiceClient.getVariable(projectId, userId).getBody();
     }
 
     @Override
     public List<DeployKey> getDeployKeys(Integer projectId, Integer userId) {
-        try {
-            return gitlabServiceClient.getDeploykeys(projectId, userId).getBody();
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
+        return gitlabServiceClient.getDeploykeys(projectId, userId).getBody();
     }
 
     @Override
     public void createDeployKey(Integer projectId, String title, String key, boolean canPush, Integer userId) {
-        try {
-            gitlabServiceClient.createDeploykey(projectId, title, key, canPush, userId);
-        } catch (FeignException e) {
-            throw new CommonException("error.deploykey.create", e);
-        }
+        gitlabServiceClient.createDeploykey(projectId, title, key, canPush, userId);
+
     }
 
     @Override
     public void addMemberIntoProject(Integer projectId, MemberDTO memberDTO) {
-        try {
             gitlabServiceClient.addMemberIntoProject(projectId, memberDTO);
-        } catch (Exception e) {
-            throw new CommonException("error.member.add", e);
-        }
     }
 
     @Override
     public void updateMemberIntoProject(Integer projectId, List<MemberDTO> list) {
-        try {
-            gitlabServiceClient.updateMemberIntoProject(projectId, list);
-        } catch (Exception e) {
-            throw new CommonException("error.member.update", e);
-        }
+        gitlabServiceClient.updateMemberIntoProject(projectId, list);
     }
 
     @Override
     public void removeMemberFromProject(Integer groupId, Integer userId) {
-        try {
             gitlabServiceClient.removeMemberFromProject(groupId, userId);
-        } catch (Exception e) {
-            throw new CommonException("error.member.remove", e);
-        }
     }
 
     @Override
     public List<GitlabProjectDO> getProjectsByUserId(Integer userId) {
-        try {
             return gitlabServiceClient.getProjectsByUserId(userId).getBody();
-        } catch (FeignException e) {
-            throw new CommonException("error.project.get.by.userId", e);
-        }
     }
 
     @Override
     public MergeRequestDO createMergeRequest(Integer projectId, String sourceBranch, String targetBranch, String title, String description, Integer userId) {
-        try {
             return gitlabServiceClient.createMergeRequest(projectId, sourceBranch, targetBranch, title, description, userId).getBody();
-        } catch (FeignException e) {
-            throw new CommonException(e);
-        }
     }
 
     @Override
